Inclusions improvements
The -b and -p parameters now work correctly with inclusions. These
need to be applied for each disjunction.

PRIORITY names (like 'err') are now converted into matchable values
(like '3').

PRIORITY inclusion matches that are not lists are treated as matches for log
levels up to the given priority, meaning these are equivalent:

- PRIORITY: 3
- PRIORITY: err
- PRIORITY: [0, 1, 2, 3]
- PRIORITY: [emerg, alert, crit, err]