Formatters no longer need to list PRIORITY in their explicit inclusions
The SelectiveReader class now takes two lists of inclusions rules, one
which will have log_level applied and one which will not. The default
rules are in the first list, and the explicit formatter filters are in
the second.

Additionally, add_disjunction() is now only called between disjuncts,
avoiding a useless one at the end.