Avoid exception on first run and no matching journal entries
When there is no existing cursor file, and the inclusions/exclusions result
in zero matching entries, an exception would be thrown because there was no
cursor value to write to the cursor file at the end of the run. To avoid this,
the cursor for the 'tail' of the journal is now obtained when there is no
cursor file (or the cursor file is empty) to use as a starting point, and if
no matching entries are found that cursor will be written to the cursor file,
which avoids the exception and also avoids scanning the same journal entries
that did not match during the next run.