Correct testing and sending of SMTP mail
An error was made when modifying `send_email` to supply the message
argument as a string (which made mock testing easier). `send_message`
requires the argument as an object, but if the message is passed as
an object then a simple equality comparison in the mock tester
will fail.

A custom matcher has been added to compare the message objects as
strings instead, so that the proper argument type can be passed to
`send_message`.