Fix SMTP mocking in Python 3.6/3.7
The `smtplib.SMTP` class in Python versions prior to 3.8 does not
handle 'unconnected' sessions silently; as a result it is necessary
to mock the `connect` and `close` methods on it to ensure that the
tests can pass. As a side effect this also avoids unintended
connections being established on the test system.