Simplify config file writing in test functions
The build_config_and_cursor fixture yields a function object which can be
called to provide a desired configuration (in either dict or string
form) to be written into the temporary config file. This eliminates
the need for the test function to explicitly write to the config file
and then flush it before the test is executed.

The function object returns the configfile and cursorfile objects,
as the fixture previously provided, so that the test function can
manipulate them when needed.

Signed-off-by: Kevin P. Fleming <kevin@km6g.us>