Use MIME-formatted email for both delivery modes
To provide consistency of configuration and output, use MIME-formatted
email for both command-based and SMTP-based delivery modes. This requires
a delivery command which can accept a fully-formatted message as input
(the usual 'mail' and 'sendmail' commands can be used in this way).

Backwards compatibility for existing configurations is preserved; non-MIME
delivery via commands is still supported, but not documented.

Signed-off-by: Kevin P. Fleming <kevin@km6g.us>